restructure folders and pass slack installation
