+++ busy_beaver/apps/github_integration/summary/task.py
@@ -1,46 +1,24 @@
-from datetime import timedelta
 import logging
 import random
 from typing import List
 
-import click
 from sqlalchemy import and_
 
-from ..blueprint import github_bp
 from .blocks import GitHubSummaryPost
 from busy_beaver.common.wrappers import SlackClient
-from busy_beaver.exceptions import ValidationError
-from busy_beaver.models import GitHubSummaryUser, SlackInstallation
-from busy_beaver.toolbox import utc_now_minus
+from busy_beaver.models import GitHubSummaryUser
 
 logger = logging.getLogger(__name__)
 
 
-@click.option("--workspace", required=True)  # Slack Workspace ID
-@github_bp.cli.command("post_github_summary", help="Post a GitHub summary")
-def post_github_summary_to_slack_cli(workspace: str):
-    boundary_dt = utc_now_minus(timedelta(days=1))
-    slack_installation = SlackInstallation.query.filter_by(
-        workspace_id=workspace
-    ).first()
-    if not slack_installation:
-        raise ValidationError("workspace not found")
-
-    # we should log that we did something somewhere
-    # also keep track of how long a summary took
-    # TODO once we are migrated over
-    fetch_github_summary_post_to_slack(slack_installation.id, boundary_dt)
-
-
-def fetch_github_summary_post_to_slack(installation_id, boundary_dt):
-    slack_installation = SlackInstallation.query.get(installation_id)
-    channel = slack_installation.github_summary_config.channel
-    slack = SlackClient(slack_installation.bot_access_token)
+def fetch_github_summary_post_to_slack(installation, boundary_dt):
+    channel = installation.github_summary_config.channel
+    slack = SlackClient(installation.bot_access_token)
 
     channel_members = slack.get_channel_members(channel)
     users: List[GitHubSummaryUser] = GitHubSummaryUser.query.filter(
         and_(
-            GitHubSummaryUser.config_id == slack_installation.github_summary_config.id,
+            GitHubSummaryUser.config_id == installation.github_summary_config.id,
             GitHubSummaryUser.slack_id.in_(channel_members),
             GitHubSummaryUser.github_username.isnot(None),
         )
